requirements-test.txt
pytest==7.4.3
pytest-cov==4.1.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
pyyaml==6.0.1
scipy==1.11.3
